final images
hey alex, please check the one that has heatmap_v02 in the name. this one I manually ran the pillars that we measured. They were in agreement with the values, which is nice. We just need to build a good last image, that shows the whole function of the 'getting the info from wire' to 'output the interface back' into a nice picture. And the work is done I believe. The final final bit, would be to run with wire itself, which i do not think we will be able to do, but you can mention it over at the conclusions.